backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-genai
python-dotenv==1.0.1
pillow==10.2.0
moviepy==1.0.3
imageio==2.33.1
requests==2.31.0
supabase==2.3.4
python-multipart